refactor: 친구 기록장 조회 서비스 로직 가독성 개선
- DTO 생성 로직은 DTO 내부로, 엔티티 조회 로직은 레포지토리로 이동
