+++ src/main/java/today/seasoning/seasoning/article/dto/FindMyFriendsArticlesResult.java
@@ -1,17 +1,20 @@
 package today.seasoning.seasoning.article.dto;
 
 import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import today.seasoning.seasoning.article.domain.Article;
 import today.seasoning.seasoning.user.dto.UserProfileResponse;
 
 @Getter
+@RequiredArgsConstructor
 public class FindMyFriendsArticlesResult {
 
-	private final UserProfileResponse profile;
-	private final FriendArticleDto article;
+    private final UserProfileResponse profile;
+    private final FriendArticleDto article;
 
-	public FindMyFriendsArticlesResult(UserProfileResponse profile,
-		FriendArticleDto article) {
-		this.profile = profile;
-		this.article = article;
-	}
+    public static FindMyFriendsArticlesResult build(Article article) {
+        return new FindMyFriendsArticlesResult(
+            UserProfileResponse.build(article.getUser()),
+            FriendArticleDto.build(article));
+    }
 }
